requirements.txt
mistralai>=0.1.0
PyPDF2>=3.0.0
pandas>=1.5.0
numpy>=1.20.0
scikit-learn>=1.0.0
python-dotenv>=1.0.0
nltk>=3.8.0
